feat: 拆分 runtimeActions.js 為領域子模組
- 拆分為 8 個領域子模組 (pageStatus, save, highlight, migration, auth, driveSync, sidepanel, diagnostics)
- 聚合層合併為單一 RUNTIME_ACTIONS frozen object，shape 完全不變
- JSDoc typedef 與 action 常量共置於同一檔案
- 根目錄 runtimeActions.js 轉為 façade
