+++ scripts/config/shared/messaging/runtime/diagnostics.js
@@ -1,0 +1,12 @@
+/**
+ * Diagnostics & Internal Actions
+ */
+
+export const DIAGNOSTICS_ACTIONS = {
+  EXPORT_DEBUG_LOGS: 'exportDebugLogs',
+  DEV_LOG_SINK: 'devLogSink',
+  DEV_LOG_SINK_BATCH: 'devLogSinkBatch',
+  PING: 'PING',
+  INIT_BUNDLE: 'INIT_BUNDLE',
+  REPLAY_BUFFERED_EVENTS: 'REPLAY_BUFFERED_EVENTS',
+};
